backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
google-generativeai==0.3.2
exa-py==1.7.1
duckduckgo-search==7.2.1
firecrawl-py==0.0.18
agno==0.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
reportlab==4.0.7
weasyprint==60.2
jinja2==3.1.2
httpx==0.25.2
aiofiles==23.2.1